coluna valor setado como bigint

+++ src/models/debtsModel.sql
@@ -3,9 +3,9 @@
     id_cliente integer not null,
     nome_cliente text not null, 
     descricao text not null,
-  	valor integer not null,
+  	valor bigint not null,
     data_vencimento date not null,    
     status text not null,    
     foreign key (nome_cliente) references clientes (nome_cliente),
     foreign key (id_cliente) references clientes (id_cliente)   
-   ); 
+   );
